refactor: collapse the .web suffixes now that there is no native sibling
With the .native files gone, the platform-extension split is ceremony: nine
files carrying a .web suffix, four one-line barrels re-exporting them, and a
resolve.extensions array in the Vite config teaching the bundler to prefer
.web over .native. All of it removed; the files now have plain names.

This also closes a real hazard in src/utils/crypto. crypto.ts and
crypto.web.ts held byte-identical copies of every named export, but because
resolve.extensions put .web.ts ahead of .ts, Vite bundled the twin while
TypeScript only ever checked crypto.ts. Nothing imported the twin directly and
nothing used its extra default export, so the two had not drifted — but
decryptUserConfig and the hex helpers were one careless edit away from a
divergence no typecheck would catch. One file now.

Renaming is all this does; no implementation changed.

Verified: tsc clean, eslint 0 errors, build succeeds, 1440 tests pass. SCSS
survived the renames — the icon-picker and user-initials selectors are present
in the built index CSS (and absent from the other chunk, so the check can
fail). The raw-name audit needed its exception paths updated, and caught the
two renames itself.
