+++ src/components/space/IconPicker/IconPicker.tsx
