+++ src/components/user/UserInitials/UserInitials.tsx
