+++ src/components/user/UserInitials/index.ts
@@ -1,3 +1,2 @@
-// Platform-aware exports (Metro/Vite will resolve .web or .native automatically)
-export { UserInitials } from './UserInitials.web';
+export { UserInitials } from './UserInitials';
 export type { UserInitialsProps } from './UserInitials.types';
